Extract ChunkManagerSource to its own file
